### 2022-10-27 #106 @whedon
[PRE REVIEW]: ProtoSyn.jl: a novel platform for computational molecular manipulation and simulation with a focus on protein design
**Submitting author:** @JosePereiraUA
**Repository:** <a href="https://github.com/sergio-santos-group/ProtoSyn.jl" target ="_blank">https://github.com/sergio-santos-group/ProtoSyn.jl</a>
**Version:**

Thanks for submitting your paper to JCON @https://github.com/JosePereiraUA. **Currently, there isn't an JCON editor assigned** to your paper.
