+++ src/Core/Methods/io-read.jl
@@ -67,7 +67,7 @@ load(::Type{T}, io::IO, ::Type{PDB}; alternative_location::String = "A", ignore_
     
     segid = atmindex = 1 # ! segment and atom index are overwritten by default 
 
-    er = r"^\w+\s+(?<aid>\d+)\s+(?<an>\w{1,4}?)\s*(?<al>(?:\w{1}(?=\w{3}\s))|\s)(?<rn>\S{3,4})\s*(?<sn>\w|[?]{0,1})\s*(?<rid>\d+)\s+(?<x>-*\d+\.\d+)\s+(?<y>-*\d+\.\d+)\s+(?<z>-*\d+\.\d+)\s+(?:\w)*\s+(?:\d+\.\d+)*\s+(?:\d+\.\d+)*\s*(?|(?:\w*\s+(\w)[0-9]*[+-]*\s*)|(?:(\w)[0-9]*[+-]*\s*)|(?<as>\w+[+-]*)(?=\s*$))$"
+    er = r"^\w+\s+(?<aid>\d+)\s+(?<an>\w{1,4}?)\s*(?<al>(?:\w{1}(?=\w{3}\s+\w{1}\s))|\s)(?<rn>\S{2,4})\s*(?<sn>[a-zA-Z]|[?]{0,1})\s*(?<rid>\d+)\s+(?<x>-*\d+\.\d+)\s+(?<y>-*\d+\.\d+)\s+(?<z>-*\d+\.\d+)\s+(?:\w)*\s+(?:\d+\.\d+)*\s+(?:\d+\.\d+)*\s*(?|(?:\w*\s+(\w)[0-9]*[+-]*\s*)|(?:(\w)[0-9]*[+-]*\s*)|(?<as>\w+[+-]*)(?=\s*$))$"
 
     aid = 0
     ignored_atoms     = Vector{Int}()
@@ -86,6 +86,8 @@ load(::Type{T}, io::IO, ::Type{PDB}; alternative_location::String = "A", ignore_
                 push!(ignored_atoms, parse(Int, atom["aid"]))
                 continue
             end
+
+            # println(atom)
 
             # * Choose alternative locations
             al = string(atom["al"])
