+++ test/generate_all_aas.jl
@@ -1,0 +1,3 @@
+using ProtoSyn
+pose = ProtoSyn.Peptides.build(ProtoSyn.Peptides.grammar, seq"ECDARGNQPMKFIHWTSLYV")
+ProtoSyn.write(pose, "all_aas.pdb")
