+++ Project.toml
@@ -12,7 +12,6 @@ CurveFit = "5a033b19-8c74-5913-a970-47c3779ef25c"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
 Downloads = "f43a241f-c20a-4ad4-852c-f6b1247861c6"
-Ghost = "4f8f7498-1303-42e1-920c-5033445536df"
 HTTP = "cd3eb016-35fb-5094-929b-558a96fad6f3"
 LightXML = "9c8b4983-aa76-5018-a973-4c85ecc9e179"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
@@ -30,4 +29,5 @@ StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+Umlaut = "92992a2b-8ce5-4a9c-bb9d-58be9a7dc841"
 YAML = "ddb6d928-2868-570f-bddf-ab3f9cf99eb6"
