+++ src/Core/Mutators/dihedral.jl
@@ -98,7 +98,8 @@ end
 function (dihedral_mutator::DihedralMutator)(pose::Pose, atoms::Vector{Atom})
     
     # DihedralMutator requires updated internal coordinates
-    ProtoSyn.c2i!(pose.state, pose.graph) # Checks pose.state.c2i flag inside
+    sync!(pose)
+    # ProtoSyn.c2i!(pose.state, pose.graph) # Checks pose.state.c2i flag inside
 
     for atom in atoms
         if rand() < dihedral_mutator.p_mut
