### 2022-11-28 6dc9e7b
pop_atom! now deletes empty residues and segments
