### 2022-12-12 #106 @odow
@JosePereiraUA you're missing the https://github.com/JuliaCon/JuliaConSubmission.jl/blob/master/paper/.latexmkrc file from the template repository.

Do you have any suggestions for reviewers?
