### 2022-12-12 50e54b5
Create .latexmkrc
