### 2022-12-12 #106 @JosePereiraUA
I've added the .latexmkrc file in [this commit](https://github.com/sergio-santos-group/ProtoSyn.jl/commit/50e54b5df01f3bb8f68d7856bc05d6fd79407da9).

From the above-mentioned list, and based on the descriptions provided, **sloede** seems to have some experience with numerical simulations. Besides that, I'd suggest anyone from the [MolSim](https://github.com/JuliaMolSim/MolSim) community.

A question: it seems the bot failed to find one of the DOI in my literature references, is there any action I have to take?

Thank you for your time!
Cheers
