### 2022-12-12 #106 @odow
@sloede and @jgreener64 are you interested in reviewing this paper for the JuliaCon proceedings?

The submission I'd like you to review is titled: "ProtoSyn.jl: a novel platform for computational molecular manipulation and simulation with a focus on protein design". You can find more information at the top of this Github issue (https://github.com/JuliaCon/proceedings-review/issues/106).

The review process at JuliaCon is based on [JOSS](https://joss.theoj.org): it takes place in a GitHub issue, is open, and author-reviewer-editor conversations are encouraged. If you have any questions please let me know.

This is the pre-review issue. After setting at least 2 reviewers we will start the review process in a seperate thread. In that thread, there will be a check list for each reviewer.

Given the lead-up to Christmas, our deadline for the first round isn't until the end of January, so there's no expectation that you'll start the review now if you are busy. I just need an indication that you're willing, otherwise I will need to find other reviewers. If you're willing, I'll ping you again in early January with a reminder.
