### 2023-03-06 #124 @whedon
[REVIEW]: ProtoSyn.jl: a novel platform for computational molecular manipulation and simulation with a focus on protein design
**Submitting author:** @https://github.com/JosePereiraUA (<a href="http://orcid.org/0000-0001-5622-8010">José Pereira</a>)
**Repository:** <a href="https://github.com/sergio-santos-group/ProtoSyn.jl" target ="_blank">https://github.com/sergio-santos-group/ProtoSyn.jl</a>
**Version:**
