### 2023-03-23 #124 @jgreener64
Protein design is an important and useful scientific problem and the authors are correct to point out the lack of variety of useable software in this area. ProtSyn.jl can therefore be a useful tool to the community, especially in combination with the steadily growing Julia ecosystem for biomolecular and atomic modelling.

The paper is fairly well written and adequately describes the software. My concerns are mainly with the packaging and testing of the software. With the below issues addressed I think this is a strong contribution and will be of interest to the protein design community in the JuliaCon Proceedings.

### Major

- I can install the software but `using ProtoSyn` and running the tests fails. I have opened this as an issue (https://github.com/sergio-santos-group/ProtoSyn.jl/issues/52).
- The build badge on the readme does not point to the actual build (https://app.travis-ci.com/github/sergio-santos-group/ProtoSyn.jl), which fails with the above error. Running CI every week, e.g. on GitHub actions, would help catch errors as they appear.
- There are no compat bounds in Project.toml, without these users may install incompatible versions of dependencies.
- There is a mismatch between the version number in Project.toml (0.4) and the version number on the GitHub release (1.1).
- Whilst it is not essential, it is conventional to register the package in the Julia registry so it can be installed with `add ProtoSyn`.

### Minor

- Briefly introduce protein design and its potential applications in the introduction of the paper.
- Protein design is currently being revolutionalised by deep learning. It is worth mentioning this at the end of the paper along with any ways that ProtoSyn.jl can be applied here.
- The paper says that "ProtoSyn.jl is completely written in Julia" but it seems it can call out to other libraries such as TorchANI, I would rephrase this.
- It might be worth citing https://doi.org/10.1016/j.cpc.2022.108452 as software that could be combined with ProtoSyn.jl.
- There are some gaps in paragraphs in the PDF of the paper, see for example the start of page 4.
- The GNU GPL 3 license of the software can be mentioned in the paper.
