### 2023-03-24 sergio-santos-group/ProtoSyn.jl#52 @mfherbst
I can reproduce the issue and get a similar stacktrace.
