### 2023-03-24 #124 @mfherbst
The authors have described a new Julia-based software platform for computational protein modelling. I resonate strongly with the author's intent to bring the promising features of Julia to this domain, which without any doubt features noteworthy computational challenges and a desperate need for more modern software. Similarly the domain of atomistic and molecular modelling is to date under-represented in the Julia ecosystem. ProtoSyn thus provides a valuabe new tool for both the Julia as well as the biomolecular communities and I overall recommend publication in JuliaCon proceedings.

Given that molecular and atomistic modelling is a huge field with many challenges, the authors make some rather bold statements in the manuscript, which come across as unbelievable. Moreover the language in the manuscript is in parts sloppy and unclear. I strongly advise to revise the manuscript, keeping these aspect in mind.

In contrast the code is well-documented and written clearly. I don't agree with the strong split-up in so many submodules, but that is personal taste. On top of the issues @jgreener64 has flagged, I have nothing to add.

# Major points
- The paper is inconsistent with respect to what knowledge is assumed. More basic aspects such as internal coordinates are explained, while "carbohydrates", "glycoproteins",  "TM-score" are aspects which are not explained or put into context. I'd suggest to keep in mind that JuliaCon proceedings target computational science as a whole   and (a) add a more lay-term introduction to protein design folding and (b) explain   terms such as the above or put them in context (e.g. is a TM-score of 0.5927 good or bad?).   This would also make it easier to anticipate why protein folding is such a tough problem   and thus performance are crucial.
- Abstract: There are a number of open-source scientific tools for molecular simulation,   that do an excellent job with respect to documentation, modularisation, software development   practices etc. [Psi4](https://psicode.org/) is such an example.   In light of that the statement "ProtoSyn.jl has the potential to modernize the way the   scientific community uses simulation tools." is hardly believable, considering scientific   computing (or even atomistic modelling) as a whole domain.
- Performance bullet point: The explanation of the "two-language problem" the authors   provide is unclear. To me the two-language problem is not that some languages are fast   and some are slow, but more about the consequences that arise when one needs to combine   two languages in a project, i.e. what is explained after the statement   "This is known as the two-language problem".
- Modular bullet point: It is not self-evident that "modular" goes hand in hand with everything   being written in one programming language. In fact I would even argue one can have   a perfectly modular code where the API of each C++ module is exposed and linked   via python (again Psi4 is a prime example).   Clearly this suffers from other issues (performance, composability, ...) due to the   two-language problem though ...
- Conclusion: The authors claim that "arbitrarily complex protocols and simulations can be easily constructed".   This statement deserves further explanation to be plausible.
- Conclusion: The authors claim "ProtoSyn.jl's development constitutes a first   attempt at a Julia-based molecular manipulation and simulation software." I   don't think this is true. There are a number of other molecular simulation
  codes in Julia out there.
- Because of https://github.com/sergio-santos-group/ProtoSyn.jl/issues/52 I have   so far not been able to run the ProtoSyn examples.

# Minor points
- Introduction: I suggest to change "namely Github" -> "such as Github" (there are other such platforms)
- Under the hood: The capitalisation is unusual here (Residues, Graph, State, ...)
- Under the hood: The frequent parenthesis make the text hard too read.
- Conclusion: I find it weird to find the computational examples in the conclusion.   Maybe have a dedicated "Examples" section. Also provide context for the examples   to the uneducated reader: Are these problems hard? How do other packages   in comparison in terms of runtime?
- The language is at times very sloppy, e.g. "new and better ways to do stuff".   The authors should revise the manuscript appropriately.
- It would be nice to cite Molly, AtomsBase, DFTK and other referenced software.
- Some of the links to ipynb files on   https://sergio-santos-group.github.io/ProtoSyn.jl/stable/getting-started/examples/
  are broken.
