### 2023-03-24 sergio-santos-group/ProtoSyn.jl#52 @JosePereiraUA
I'll take a look at this issue now.
