### 2023-03-24 sergio-santos-group/ProtoSyn.jl#52 @JosePereiraUA
The [dev ](https://github.com/sergio-santos-group/ProtoSyn.jl/tree/dev)branch was working ok, I just forgot to merge the changes to the master branch. Commit https://github.com/sergio-santos-group/ProtoSyn.jl/commit/4a201aa23a77c4297bfc9813ef8b9a59797643f4 fixes this, the current master branch should import just fine. Let me know if you still have troubles.
