### 2023-03-24 #124 @JosePereiraUA
Thank you so much for the constructive feedback.
I can certainly revise the manuscript if that is necessary, smoothing out the claims. I would like to take the opportunity to explain that ProtoSyn.jl was born almost 4 years ago, when some (if not most) of the claims made "more" sense. Fortunately, the field has exploded in the right direction and indeed a revision on how "revolutionary" ProtoSyn.jl is, now, is probably advised.

I would also like to add I've fixed the above-mentioned issue (see https://github.com/sergio-santos-group/ProtoSyn.jl/issues/52#issuecomment-1483353303). Any tips on the next steps?
